more revocation fixes; align API with recent updates
